% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{aggregate_matrix}
\alias{aggregate_matrix}
\title{Compute Group Means by Cluster Membership}
\usage{
aggregate_matrix(X, v)
}
\arguments{
\item{X}{A numeric matrix of size N x P.}

\item{v}{A vector of group membership indicators of length N.}
}
\value{
A K x P matrix of group means, where K is the number of unique values in \code{v}.
}
\description{
Computes column means of a numeric matrix grouped by a vector of categorical indicators.
}
\keyword{internal}
