% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{epa_clustered_known}
\alias{epa_clustered_known}
\title{Clustered EPA (C-EPA) Test with Known Group Memberships}
\usage{
epa_clustered_known(Z, id, time, gamma, lrv = "EWC", lrv_par = NULL)
}
\arguments{
\item{Z}{Matrix (NT x P) of loss differentials or moment conditions.}

\item{id}{Vector of length NT for unit identifiers.}

\item{time}{Vector of length NT for time identifiers.}

\item{gamma}{Vector of length N (number of units), giving cluster assignments (values in {1, ..., K}).}

\item{lrv}{Character; "EWC" (default) or "NeweyWest".}

\item{lrv_par}{Integer; lag (NeweyWest) or number of spectral windows (EWC).}
}
\value{
A list with elements:
  \describe{
    \item{W}{Wald statistic}
    \item{pval}{P-value}
  }
}
\description{
Clustered EPA (C-EPA) Test with Known Group Memberships
}
