% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{epa_clustered_split}
\alias{epa_clustered_split}
\title{Split-Sample Clustered EPA}
\usage{
epa_clustered_split(
  df,
  id,
  time,
  Z_names,
  K,
  prop = 0.5,
  method = "EWC",
  param = NULL,
  Ninit = 10,
  iter.max = 10
)
}
\arguments{
\item{df}{Balanced panel data.frame}

\item{id}{Character; unit identifier column name}

\item{time}{Character; time identifier column name}

\item{Z_names}{Character vector; variable names for clustering}

\item{K}{Integer; number of clusters for KMeans}

\item{prop}{Numeric; proportion of time periods for R (training) sample}

\item{method}{Character; "EWC" or "NeweyWest" for LRV estimator}

\item{param}{Integer; lag (NW) or B (EWC)}

\item{Ninit}{Integer; number of KMeans initializations}

\item{iter.max}{Integer; max iterations in KMeans}
}
\value{
List with cluster assignments and test results from P
}
\description{
Split-Sample Clustered EPA
}
