% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{split_panel_matrix}
\alias{split_panel_matrix}
\title{Split a balanced panel matrix into estimation (R) and evaluation (P) samples}
\usage{
split_panel_matrix(Z, id, time, prop = 0.5, maxlag = 0)
}
\arguments{
\item{Z}{Matrix of dimension (N*T) x K}

\item{id}{Numeric vector of length N*T (unit id)}

\item{time}{Numeric vector of length N*T (time id)}

\item{prop}{Proportion of time periods for R sample}

\item{maxlag}{Lags to remove from R sample}
}
\value{
List with matrices and aligned id/time: Z_R, Z_P, id_R, id_P, time_R, time_P, P, Tobs
}
\description{
Split a balanced panel matrix into estimation (R) and evaluation (P) samples
}
\keyword{internal}
